AxisType: Deprecated isXY and isSpatial
Fundamentally, AxisTypes should not be defined by their label. The label
uniquely identifies and distinguishes axes and these methods are tied to
a 5D model, where X and Y are significant and XYZ are spatial.

To move to an N-D model, what really matters is the order of the axes
and the number of planar axes. Thus isXY and isSpatial would become
unnecessary.

These methods are thus deprecated and when conversion of consuming
(SCIFIO) code to an N-D model is complete, and propagated downstream
(e.g. to IJ2), these methods can be removed.