CalibratedSpace: reinstate calibration API
For backwards compatibility, we should cut at least one release with
these methods still in place, but deprecated. Then afterward we can
remove the deprecated API. Otherwise, upgrading downstream code to new
versions of ImgLib2 is greatly complicated by the fact that API was
simply removed from release to release, with no deprecation period.

Keeping the methods available for one release allows downstream code to
upgrade to the new API in phases without breaking its build. Otherwise,
the entire upgrade must be done as a single monster commit, which is
difficult to code review and cannot later be bisected for debugging
purposes.