Rename UncalibratedAxis to IdentityAxis
Rather than thinking of this implementation as "uncalibrated" (which
would simply be a TypedAxis that does not implement CalibratedAxis at
all), let's repurpose this class as the identity function (y = x) since
that's already what it was in practice. The only difference conceptually
is that it is fine for an IdentityAxis to have a non-null unit; this
actually makes the implementation more concise, so no issue there.