WIP: CalibratedAxis: use copy method instead of update
The update approach is cool but, as implemented, is fixed in nature: it
can perform some potentially useful axis conversions but they are all
hardcoded. Since we have no code downstream that specifically needs the
power of such conversions, let's instead punt for now and merely provide
each axis with the ability to clone itself via a copy() method.

For the copy() method, we avoid the use of recursive generics, instead
opting to always override copy() with a fully narrowed return type.
Since the axis type hierarchy is rather flat right now, this works fine.