PolynomialAxis: remove coefficient count limit
Firstly, it may be desirable to first create the axis before
programmatically adding a variable number of coefficients which are not
not known in advance.

Secondly, technically a polynomial axis can have less than three
coefficients. I think this check was just in place because LinearAxis
serves the same purpose if there are only one or two coefficients.