Clean up dependencies
In particular, do not let managed imglib2.version take precedence.
This updates the parent to pom-scijava 1.161.

Rather, each componenet should depend on the current working copy's
version of each module in the multi-module build (i.e., on
${project.version}, not the ${imglib2.version} declared in pom-scijava).

However, managed versions always take precedence. So it is not enough to
declare '<version>${project.version}</version>' as that will be
overridden by the dependencyManagement of pom-scijava. Instead, we
override the ${imglib2.version} property to equal ${project.version} so
that the current multi-module build's version *is* the managed version.
This has the nice side effect of letting us remove the '<version>' tags
altogether from all project dependencies.