AbstractImageDisplayViewer: add "(V)" trailer
When the image is a "virtual" one (i.e., a SCIFIO cell image), we put a
"(V)" in the title bar. This is a holdover from ImageJ1, and it would be
good to do something more sensible in the future, but for now, as a
simple indicator, it is better than nothing.

This logic was formerly (erroneously) in AbstractDisplayViewer.