DatasetService: make save methods return Metadata
When SCIFIO writes a dataset, the relevant SCIFIO Writer constructs a
Metadata descriptor. It is nice to be able to access this descriptor
after the fact, in case the Writer recorded any details of the
transaction there. For example, the scifio-omero project introduces
an OMEROFormat whose Writer records the ID of the new OMERO Image.