OverlayService: fix addOverlays method signature
The typing of the list of Overlays should be List<? extends Overlay>.
This is in line with the rule of thumb for good use of generics in APIs:
be liberal in what you accept ("? extends" for method arguments) and
precise in what you produce (exact typing in method return values).