Avoid overriding ImageJ 1.x' File>New>Image...
ImageJ 1.x has a feature (which could also be called a 'design
limitation') that the menu labels of commands are used to identify
commands in macros. That is convenient just until the point when you
want to add, say, a plugin into the menu path File>Import>Image... and
that plugin clashes with the already installed File>New>Image...

Let's play nice with ImageJ 1.x (as if we ever did anything remotely
different...) and just rename our import/export plugins' menu paths
using the common work-around: add spaces.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>