Let the MediaWiki client default to the ImageJ Wiki
The ImageJ Wiki -- http://wiki.imagej.net/ -- is just another skin and
default page for the Fiji Wiki, to reflect the nature of the true
relationship between the Fiji and the ImageJ2 project: very, very
collaborative and collegial.

But users want to see the name "ImageJ" and not "Fiji". Well, let them
have it!

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>