DefaultTableDisplay: handle double array as table
In this specific case (double[] and double[][]), it is easy to wrap the
data in DoubleColumn objects. Ideally, we want to support all types of
primitive 1D and 2D arrays, or even all Collection<N extends Number>,
but that will require a little more infrastructural work to accomplish.
This approach is good enough for now, for MATLAB matrix support, which
always uses doubles.