DefaultOverlayService: eliminate PointSet usage
The RoiPointSet is cool, allowing you to express a RegionOfInterest
(which is a RealRandomAccessible) as an IterableInterval. But actually,
we can do the same thing using Views, via the raster, interval and
iterable methods.

Admittedly, it would be nice if Views had one or two more built-in
methods to facilitate this (particularly computing the Interval's
min/max bounds). Perhaps we can contribute that logic upstream later.