POM: remove no-longer-needed SCIFIO dependency
The io.scif:scifio project will now depend on imagej-common, instead
of the other way around. The rationale is that the ImageJ library
(net.imagej:imagej-common) is the place the adds metadata richness on
top of the ImgLib2 library (net.imglib2:imglib2). And the SCIFIO
library (io.scif:scifio) is built on top of that metadata richness.

For details, see:
http://imagej.net/Architecture