DatasetService: deprecate almost-obsolete methods
This deprecation is in preparation for the removal of the SCIFIO
dependency from imagej-common (the dependency is being inverted).

Ultimately, logic for opening up Datasets needs to live in SCIFIO, not
imagej-common. (And alternately, a more agnostic way of opening and
saving Datasets can be had using the org.scijava.io.IOService.)