Split ActiveImagePreprocessor into modular pieces
This change is good from a modularity and extensibility standpoint, and
addresses a long-standing TODO.

However, there is another huge win: it also avoids a recently introduced
problem of the legacy layer, where calling
ImageDisplayService#getActiveImageDisplay() would trigger a potentially
expensive IJ1->IJ2 sync operation of the active ImagePlus. The new
structure waits to obtain the active ImageDisplay, or any derivative
thereof, until _after_ it is confirmed that a single input of compatible
type exists and wants to be populated.

For details and discussion, see:
http://imagej.1557.x6.nabble.com/Running-scripts-with-an-open-Virtual-Hyperstack-tp5012163p5012169.html