Undeprecate some lynchpin classes
The ImgPlus and friends are in heavy use in ImageJ Ops and elsewhere,
and there is basically no choice right now but to use them, unless you
want to stick with ImgLib2 classes (which lack crucial metadata) or
ImageJ 1.x classes (which have various limitations).

So we change the wording accordingly, to avoid the boatload of
obnoxious and unfixable deprecation warnings, in favor of a bold
notice in the class javadoc, which no one will read.

Eventually (hopefully by end of spring 2016), we will have something
better, and _then_ we can deprecate these classes.