Generalize SingleInputPreprocessor
The SingleInputPreprocessor can now lean on the ConvertService to find
all types that are compatible withe the preprocessor's generic type, and
then resolve any single unresolved input of these types.

This allows a simplification of existing SingleInputPreprocessors (in
this case, ActiveDatasetPreprocessor) as they can now satisfy module
inputs of a range of types, and only need to contain the logic for
obtaining an original instance.