ImgPlus: default name is now null
After discussion with ctrueden, we agree that there is no contract that
a Named object can not have a null name. This was inconsistently
enforced in ImgPlus, where the constructor forced the name to be
non-null, but the name could be nulled out by setName.

This inconsistency is resolved by removing the default name and allowing
name to be null if it is not explicitly set. This is important to allow
consumers to know when to set a default name.