ResultsPostprocessor: handle Object outputs better
If the output type is Object, let's look at the actual object value when
deciding whether to lump it into the table. We use this heuristic here
because an Object output is very often the default "result" output, and
we want that value to show in the table if its type is compatible.