Improve the Travis CI build process
The Travis build script is now curled from the scijava/scijava-scripts
repository, so that improvements to the build process can be pushed
centrally and all interested repositories benefit automatically.

Release tags of the form <foo>-[0-9].<bar> are now greenlit for build,
so that release tags of that form are deployed successfully.

The ImageJ Jenkins is no longer pinged in response to the build, since
there were no concrete use cases for doing so, there are unlikely to be
in the future, and removing this logic simplifies the configuration.