AxesTest: relax Axes.knownTypes() assumptions
The number of known axis types will depend on whether Axes.get(...) has
been called to lookup and construct new types on demand. When running
the unit suite, other tests are doing this, thus "polluting" the set of
axis types. Let's be tolerant of this, and merely verify that the five
core types (X, Y, Z, C, T) are accounted for.