ImageDisplayService: add type-safe wrapper methods
* createDatasetView: Dataset -> DatasetView
* createImageDisplay: Data -> ImageDisplay
* createImageDisplay: DataView -> ImageDisplay

For createDatasetView, the createDataView method already did this, but
there was no runtime guarantee that the wrapped DataView would implement
DatasetView, and a rogue DataView plugin implementation could
potentially hose it up.

For createImageDisplay, the DisplayService's createDisplay method
already did this, but again, no runtime guarantee of ImageDisplay on the
return type.

These new methods are more convenient for callers to avoid the need to
recapitulate instanceof checks. They also hopefully make it slightly
more clear that any DataView wrapping a Dataset needs to implement
DatasetView, and any Display wrapping a DataView needs to implement
ImageDisplay.