Simplify converter plugin implementations
They now extend a common base class ConciseConverter, which provides
a constructor signature to declare everything that matters.

And fix a problem with the ROITreeToMaskPredicateConverter.
It was checking destination type assignability in the wrong direction.
It wants to be an instance-sensitive converter, allowing conversion of
ROITree to SomeMaskPredicateSubclass as long as that particular ROITree
contains exactly one child assignable to SomeMaskPredicateSubclass.
But in order to accomplish that, we need to avoid falling back to any
instance-insensitive canConvert implementations, because by types alone,
we cannot know that a ROITree is convertible to that type, and hence
must return false to that query.