Add converters for all common image types
This work is a major step toward resolving imagej/imagej-legacy#229,
but does not quite do the entire job, since ij.ImagePlus is another
image type on that list, which is not available from imagej-common.
But we can add converters to imagej-legacy for ImagePlus to+from all
other image types, and then we'll have everything covered.

See also imagej/imagej-ops#54.

Note that these new converters triggered one failing test,
SingleInputPreprocessorTest#testSingleImageDisplay, because the
ActiveDatasetPreprocessor was getting a chance to fill the single
ImageDisplay parameter before the ActiveImageDisplayPreprocessor did,
and so the ActiveDatasetPreprocessor would wrap the active Dataset into
a *new ImageDisplay* via the newly added DatasetToImageDisplayConverter.
To avoid this issue, this commit bumps up the priorities of certain
SingleInputPreprocessors: ActiveImageDisplayPreprocessor highest, then
ActiveDatasetViewPreprocessor, then ActiveDataViewPreprocessor, then
ActiveDatasetPreprocessor. This adjustment results in all the tests
passing, but we will need to continue to be careful about the
SingleInputPreprocessor ordering. As things stand, there is probably
still some surprising behavior under the hood, because now when there is
a single Dataset input, it will be the ActiveImageDisplayPreprocessor
that fills it by unwrapping that Dataset from itself via the
ImageDisplayToDatasetConverter. We could consider removing a bunch of
no-longer-necessary SingleInputPreprocessor implementations, but for
now, I am choosing not to rock the boat further.